requirements.txt
black
flake8
isort
loguru
mkdocs
pip
python-dotenv
tqdm
typer

pandas
numpy
matplotlib
h5py
nexusformat
pickle

-e .
